chore: include default PR template for web use.
As mentioned in the conversation on issue #58, the previously added PR
templates in the `.github/PULL_REQUEST_TEMPLATE` directory were only
accessible via a query, not from the GitHub Web App. To remedy this, a
new `.github/pull_request_template.md` file is created which compiles
all of the existing PR templates into a single file for easy access from
the GitHub Web App.

Closes #58.