63 task fix issue forms (#64)
* fix: adjust issue types to new names

The issue types were renamed to include a number for ordering purposes.
This needed to be reflected in the issue forms for the correct creation
of issue types.

* fix: include missin user role to US issue form

The _Administrator_ role was missing for the User Story role
definitions. Hence the User Stories created with this issue form were
unable to select this role.