docs: include state diagram in `README.md` (#78)
* docs: include state diagram in `README.md`

Include a state diagram using mermaid which displays the whole process of a travel request application and the states that such application can be in, as well as the states from which a different one can be achieved.

* docs: include missing connections and descriptions

Include the missing connections, and add a description to all relevant links between states. Also utilize the `<<choice>>` and `<<join>>` states to better reflect how the state flow goes.