docs(contributing): release merge workflow
Due to some lack of foresight by me, the squash and rebase workflow for
releases meant that the `main` and the `development` branch would always
have merge conflicts due to having divergent histories.

This issue was outlined by @Porto1090
<165714406+Porto1090@users.noreply.github.com>, but was only addressed
at a later date once I figured out what was wrong and fixed the
previously squash and rebased release and turned it into a merge
workflow instead.

This is then the documentation change to reflect such fact.